Fix chord note frequency calculation type casting

+++ lib/models/chord.dart
@@ -60,7 +60,9 @@ class Chord {
   ChordNote _buildNoteForString(int stringIndex, int fret) {
     final _StringTuning tuning = _stringTunings[stringIndex];
     final int midi = tuning.midi + fret;
-    final double frequency = 440 * pow(2, (midi - 69) / 12);
+    final double frequency = 440.0 * pow(2, (midi - 69) / 12).toDouble();
+
+
     return ChordNote(
       stringIndex: stringIndex,
       midi: midi,
